+++ lucene/core/src/java/module-info.java
@@ -64,8 +64,6 @@ module org.apache.lucene.core {
 
   exports org.apache.lucene.util.quantization;
   exports org.apache.lucene.codecs.hnsw;
-  exports org.apache.lucene.internal.vectorization to
-      org.apache.lucene.benchmark.jmh;
 
   provides org.apache.lucene.analysis.TokenizerFactory with
       org.apache.lucene.analysis.standard.StandardTokenizerFactory;
